requirements.txt
libimagequant~=2.17.0
basicsr~=1.4.2
diffusers~=0.14.0
realesrgan~=0.3.0
transformers~=4.24.0
imageio==2.27.0
lazy_loader==0.2
networkx==3.1
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
# pkg_resources==0.0.0
PyWavelets==1.4.1
scikit-image==0.20.0
scipy==1.9.1
sklearn==0.0.post4
tifffile==2023.4.12
torch==2.0.0
pytorch-msssim==0.2.1
tqdm==4.65.0
lpips==0.1.4
einops==0.6.1
timm==0.9.2